[AMQCPP-30] updated output of priority in toString()
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@508908 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/connector/openwire/commands/ConsumerInfo.cpp
@@ -162,7 +162,7 @@ std::string ConsumerInfo::toString() const {
     stream << " Value of NoLocal = " << this->isNoLocal() << std::endl;
     stream << " Value of Exclusive = " << this->isExclusive() << std::endl;
     stream << " Value of Retroactive = " << this->isRetroactive() << std::endl;
-    stream << " Value of Priority = " << this->getPriority() << std::endl;
+    stream << " Value of Priority = " << (int)this->getPriority() << std::endl;
     for( size_t ibrokerPath = 0; ibrokerPath < this->getBrokerPath().size(); ++ibrokerPath ) {
         stream << " Value of BrokerPath[" << ibrokerPath << "] is Below:" << std::endl;
         if( this->getBrokerPath()[ibrokerPath] != NULL ) {
